+++ src/Plutarch/Extra/Ord.hs
@@ -344,22 +344,22 @@ pisSortedBy ::
     (PElemConstraint ell a, PListLike ell) =>
     Term s (PComparator a :--> ell a :--> PBool)
 pisSortedBy = phoistAcyclic $
-    plam $ \cmp xs ->
-        let success = pconstant True
-         in precList2 (go cmp) success # xs
+    plam $ \cmp ->
+        precListLookahead (go cmp) (const (pconstant True)) (pconstant True)
   where
     go ::
         forall (s' :: S).
         Term s' (PComparator a) ->
+        Term s' (a :--> ell a :--> PBool) ->
         Term s' a ->
         Term s' a ->
-        Term s' PBool ->
+        Term s' (ell a) ->
         Term s' PBool
-    go cmp x y result =
+    go cmp self h peek rest =
         pif
-            result
-            (pif (pleqBy # cmp # x # y) result (pconstant False))
-            result
+            (pleqBy # cmp # h # peek)
+            (self # peek # rest)
+            (pconstant False)
 
 {- | Verifies that a list-like structure is both ordered (by the 'POrd' instance
  it's full of) and has no duplicates (by the 'PEq' instance it's full of).
@@ -387,23 +387,23 @@ pallUniqueBy ::
     (PElemConstraint ell a, PListLike ell) =>
     Term s (PComparator a :--> ell a :--> PMaybe PBool)
 pallUniqueBy = phoistAcyclic $
-    plam $ \cmp xs ->
-        let success = pcon . PJust . pconstant $ True
-         in precList2 (go cmp) success # xs
+    plam $ \cmp ->
+        precListLookahead (go cmp) (const success) success
   where
+    success :: forall (s' :: S). Term s' (PMaybe PBool)
+    success = pcon . PJust . pconstant $ True
     go ::
         forall (s' :: S).
         Term s' (PComparator a) ->
+        Term s' (a :--> ell a :--> PMaybe PBool) ->
         Term s' a ->
         Term s' a ->
-        Term s' (PMaybe PBool) ->
+        Term s' (ell a) ->
         Term s' (PMaybe PBool)
-    go cmp x y result = pmatch result $ \case
-        PNothing -> result
-        PJust _ -> pmatch (pcompareBy # cmp # x # y) $ \case
-            PLT -> result
-            PEQ -> pcon . PJust . pconstant $ False
-            PGT -> pcon PNothing
+    go cmp self h peek t = pmatch (pcompareBy # cmp # h # peek) $ \case
+        PLT -> self # peek # t
+        PEQ -> pcon . PJust . pconstant $ False
+        PGT -> pcon PNothing
 
 {- | As 'pallUnique', but errors if the list-like argument is found to be
  unsorted instead of returning 'PNothing'.
@@ -468,14 +468,14 @@ ptryMergeBy = phoistAcyclic $
         Term s' (ell a) ->
         Term s' (ell a)
     go cmp self x xs y ys = pmatch (pcompareBy # cmp # x # y) $ \case
-        PGT -> phandleList ys (pcons # y #$ passertSorted # cmp # xs) $ \y' ys' ->
+        PGT -> pcons # y #$ phandleList ys (passertSortedLookahead # cmp # x # xs) $ \y' ys' ->
             pmatch (pcompareBy # cmp # y # y') $ \case
-                PGT -> ptraceError "ptryMergeBy: argument list-like out of order"
-                _ -> pcons # y #$ self # x # xs # y' # ys'
-        _ -> phandleList xs (pcons # x #$ passertSorted # cmp # ys) $ \x' xs' ->
+                PGT -> unorderedError
+                _ -> self # x # xs # y' # ys'
+        _ -> pcons # x #$ phandleList xs (passertSortedLookahead # cmp # y # ys) $ \x' xs' ->
             pmatch (pcompareBy # cmp # x # x') $ \case
-                PGT -> ptraceError "ptryMergeBy: argument list-like out of order"
-                _ -> pcons # x #$ self # x' # xs' # y # ys
+                PGT -> unorderedError
+                _ -> self # x' # xs' # y # ys
 
 {- | Sort a list-like by the 'POrd' instance of its contents.
 
@@ -507,7 +507,7 @@ psortBy = phoistAcyclic $
             cmp
             (const (pconvertLists #))
             pmergeUnsafe
-            #$ pmergesort
+            #$ pmergeStart_2_3_4
             # cmp
             # xs
 
@@ -537,7 +537,7 @@ pnubSortBy = phoistAcyclic $
             cmp
             (\cmp' xs' -> pconvertLists #$ pnubUnsafe # cmp' # xs')
             pmergeUnsafeNoDupes
-            #$ pmergesort
+            #$ pmergeStart_2_3_4
             # cmp
             # xs
 
@@ -558,10 +558,10 @@ pmergeAllUnsafe ::
 pmergeAllUnsafe cmp whenSingleton whenMerging = pfix #$ plam $ \self xs ->
     phandleList xs pnil $ \x' xs' ->
         phandleList xs' (whenSingleton cmp x') $ \_ _ ->
-            self #$ go # xs
+            self #$ pmergePairs # xs
   where
-    go :: Term s (PList (PList a) :--> PList (PList a))
-    go = pfix #$ plam $ \self xs ->
+    pmergePairs :: Term s (PList (PList a) :--> PList (PList a))
+    pmergePairs = pfix #$ plam $ \self xs ->
         phandleList xs pnil $ \x' xs' ->
             phandleList xs' xs $ \x'' xs'' ->
                 pcons # whenMerging cmp x' x'' #$ self # xs''
@@ -578,22 +578,19 @@ pmergeUnsafe ::
     Term s (PList a)
 pmergeUnsafe cmp xs ys =
     phandleList xs ys $ \x' xs' ->
-        phandleList ys xs $ \y' ys' ->
-            (pfix #$ plam $ go) # x' # xs' # y' # ys'
+        (pfix #$ plam $ go) # x' # xs' # ys
   where
     go ::
-        Term s (a :--> PList a :--> a :--> PList a :--> PList a) ->
+        Term s (a :--> PList a :--> PList a :--> PList a) ->
         Term s a ->
         Term s (PList a) ->
-        Term s a ->
         Term s (PList a) ->
         Term s (PList a)
-    go self xsHead xsTail ysHead ysTail =
-        pmatch (pcompareBy # cmp # xsHead # ysHead) $ \case
-            PGT -> pcons # ysHead #$ phandleList ysTail (pcons # xsHead # xsTail) $
-                \newYsHead newYsTail -> self # xsHead # xsTail # newYsHead # newYsTail
-            _ -> pcons # xsHead #$ phandleList xsTail (pcons # ysHead # ysTail) $
-                \newXsHead newXsTail -> self # newXsHead # newXsTail # ysHead # ysTail
+    go self x' xs' ys' =
+        phandleList ys' (pcons # x' # xs') $ \y'' ys'' ->
+            pmatch (pcompareBy # cmp # x' # y'') $ \case
+                PGT -> pcons # y'' #$ self # x' # xs' # ys''
+                _ -> pcons # x' #$ self # y'' # ys'' # xs'
 
 -- Merges two PLists, throwing out duplicates.
 --
@@ -607,25 +604,20 @@ pmergeUnsafeNoDupes ::
     Term s (PList a)
 pmergeUnsafeNoDupes cmp xs ys =
     phandleList xs (pnubUnsafe # cmp # ys) $ \x' xs' ->
-        phandleList ys (pnubUnsafe # cmp # xs) $ \y' ys' ->
-            (pfix #$ plam $ go) # x' # xs' # y' # ys'
+        (pfix #$ plam $ go) # x' # xs' # ys
   where
     go ::
-        Term s (a :--> PList a :--> a :--> PList a :--> PList a) ->
+        Term s (a :--> PList a :--> PList a :--> PList a) ->
         Term s a ->
         Term s (PList a) ->
-        Term s a ->
         Term s (PList a) ->
         Term s (PList a)
-    go self xsHead xsTail ysHead ysTail =
-        pmatch (pcompareBy # cmp # xsHead # ysHead) $ \case
-            PLT -> pcons # xsHead #$ phandleList xsTail (pnubUnsafe # cmp #$ pcons # ysHead # ysTail) $
-                \newXsHead newXsTail -> self # newXsHead # newXsTail # ysHead # ysTail
-            PEQ -> pcons # xsHead #$ phandleList xsTail (pnubUnsafe # cmp # ysTail) $
-                \newXsHead newXsTail -> phandleList ysTail (pnubUnsafe # cmp # xsTail) $
-                    \newYsHead newYsTail -> self # newXsHead # newXsTail # newYsHead # newYsTail
-            PGT -> pcons # ysHead #$ phandleList ysTail (pnubUnsafe # cmp #$ pcons # xsHead # xsTail) $
-                \newYsHead newYsTail -> self # xsHead # xsTail # newYsHead # newYsTail
+    go self x' xs' ys' =
+        phandleList ys' (pnubUnsafe # cmp #$ pcons # x' # xs') $ \y'' ys'' ->
+            pmatch (pcompareBy # cmp # x' # y'') $ \case
+                PLT -> pcons # x' #$ self # y'' # ys'' # xs'
+                PEQ -> self # y'' # ys'' # xs'
+                PGT -> pcons # y'' #$ self # x' # xs' # ys''
 
 -- Removes all duplicates from a sorted list.
 --
@@ -634,28 +626,32 @@ pmergeUnsafeNoDupes cmp xs ys =
 pnubUnsafe ::
     forall (a :: S -> Type) (s :: S).
     Term s (PComparator a :--> PList a :--> PList a)
-pnubUnsafe = phoistAcyclic $ plam $ \cmp -> precList2 (go cmp) pnil
+pnubUnsafe = phoistAcyclic $
+    plam $ \cmp ->
+        precListLookahead (go cmp) (psingleton #) pnil
   where
     go ::
         forall (s' :: S).
         Term s' (PComparator a) ->
+        Term s' (a :--> PList a :--> PList a) ->
         Term s' a ->
         Term s' a ->
         Term s' (PList a) ->
         Term s' (PList a)
-    go cmp x peek rest =
+    go cmp self h peek t =
         pif
-            (pequateBy # cmp # x # peek)
-            rest
-            (pcons # x # rest)
+            (pequateBy # cmp # h # peek)
+            (self # peek # t)
+            (pcons # h #$ self # peek # t)
 
--- Breaks the argument into sorted chunks; currently, these go up to size 4. To
--- speed up the chunk sorting, we use sorting networks.
-pmergesort ::
+-- Breaks the argument into sorted chunks; currently, these range in size from 2
+-- to 4, with preference for larger chunks. To speed up the chunk sorting, we
+-- use sorting networks.
+pmergeStart_2_3_4 ::
     forall (a :: S -> Type) (ell :: (S -> Type) -> S -> Type) (s :: S).
     (PElemConstraint ell a, PListLike ell) =>
     Term s (PComparator a :--> ell a :--> PList (PList a))
-pmergesort = phoistAcyclic $
+pmergeStart_2_3_4 = phoistAcyclic $
     pfix #$ plam $ \self cmp ->
         pmatchList pnil $
             \_0 -> pmatchList (plist [psing _0]) $
@@ -761,24 +757,25 @@ pswap ::
     Term s r
 pswap cmp x y cont = pif (pleqBy # cmp # x # y) (cont x y) (cont y x)
 
--- Similar to zipWith f xs (tail xs)
-precList2 ::
+-- precList with 'lookahead' capabilities
+precListLookahead ::
     forall (a :: S -> Type) (r :: S -> Type) (ell :: (S -> Type) -> S -> Type) (s :: S).
     (PElemConstraint ell a, PListLike ell) =>
-    (Term s a -> Term s a -> Term s r -> Term s r) ->
+    (Term s (a :--> ell a :--> r) -> Term s a -> Term s a -> Term s (ell a) -> Term s r) ->
+    (Term s a -> Term s r) ->
     Term s r ->
     Term s (ell a :--> r)
-precList2 whenContinuing whenDone = plam $ \xs ->
-    phandleList xs whenDone $ \x xs' ->
-        (pfix #$ plam $ go) # x # xs'
+precListLookahead whenContinuing whenOne whenDone = plam $ \xs ->
+    phandleList xs whenDone $ \x' xs' ->
+        (pfix #$ plam $ go) # x' # xs'
   where
     go ::
         Term s (a :--> ell a :--> r) ->
         Term s a ->
         Term s (ell a) ->
         Term s r
-    go self x xs = phandleList xs whenDone $ \y ys ->
-        whenContinuing x y (self # y # ys)
+    go self h t = phandleList t (whenOne h) $ \peek rest ->
+        whenContinuing self h peek rest
 
 -- pelimList with the list-like first, and handles the 'nil case' before the
 -- 'cons' case
@@ -798,7 +795,25 @@ passertSorted ::
     Term s (PComparator a :--> ell a :--> ell a)
 passertSorted = phoistAcyclic $
     plam $ \cmp xs ->
+        phandleList xs xs $ \x' xs' ->
+            passertSortedLookahead # cmp # x' # xs'
+
+-- as passertSorted, but with the 'lookahead' already done
+passertSortedLookahead ::
+    forall (a :: S -> Type) (ell :: (S -> Type) -> S -> Type) (s :: S).
+    (PElemConstraint ell a, PListLike ell) =>
+    Term s (PComparator a :--> a :--> ell a :--> ell a)
+passertSortedLookahead = phoistAcyclic $
+    plam $ \cmp x xs ->
         pif
             (pisSortedBy # cmp # xs)
-            xs
-            (ptraceError "ptryMergeBy: argument list-like out of order")
+            ( phandleList xs (psingleton # x) $ \x' _ ->
+                pif
+                    (pleqBy # cmp # x # x')
+                    (pcons # x # xs)
+                    unorderedError
+            )
+            unorderedError
+
+unorderedError :: forall (a :: S -> Type) (s :: S). Term s a
+unorderedError = ptraceError "ptryMergeBy: argument list-like out of order"